Use sealed types for hprof record types and internal heap dump elements.
This makes it possible to have the compiler check that consuming code
is handling every possible subtype.
