--- src/main/java/org/adoptopenjdk/jheappo/heapdump/CPUSamples.kt
@@ -1,16 +1,0 @@
-package org.adoptopenjdk.jheappo.heapdump
-
-/*
- * Copyright (c) 2018 Kirk Pepperdine.
- * Licensed under https://github.com/AdoptOpenJDK/jheappo/blob/master/LICENSE
- * Instructions: https://github.com/AdoptOpenJDK/jheappo/wiki
- */
-
-
-import org.adoptopenjdk.jheappo.io.HeapProfileRecord
-
-class CPUSamples(body: EncodedChunk) : HeapProfileRecord() {
-    companion object {
-        const val TAG: UByte = 0x0DU
-    }
-}
