--- src/main/java/org/adoptopenjdk/jheappo/objects/RootNativeStack.kt
@@ -1,18 +1,0 @@
-package org.adoptopenjdk.jheappo.objects
-
-/*
- * Copyright (c) 2018 Kirk Pepperdine.
- * Licensed under https://github.com/AdoptOpenJDK/jheappo/blob/master/LICENSE
- * Instructions: https://github.com/AdoptOpenJDK/jheappo/wiki
- */
-
-import org.adoptopenjdk.jheappo.heapdump.EncodedChunk
-
-class RootNativeStack(buffer: EncodedChunk) : HeapObject(buffer) {
-
-    companion object {
-        const val TAG: UByte = 0x04U
-    }
-
-    private val threadSerialNumber: UInt = buffer.extractU4()
-}
