--- src/main/java/org/adoptopenjdk/jheappo/objects/RootStickyClass.kt
@@ -1,15 +1,0 @@
-package org.adoptopenjdk.jheappo.objects
-
-/*
- * Copyright (c) 2018 Kirk Pepperdine.
- * Licensed under https://github.com/AdoptOpenJDK/jheappo/blob/master/LICENSE
- * Instructions: https://github.com/AdoptOpenJDK/jheappo/wiki
- */
-
-import org.adoptopenjdk.jheappo.heapdump.EncodedChunk
-
-class RootStickyClass(buffer: EncodedChunk) : HeapObject(buffer) {
-    companion object {
-        const val TAG: UByte = 0x05U
-    }
-}
