--- src/main/java/org/adoptopenjdk/jheappo/objects/RootThreadBlock.kt
@@ -1,22 +1,0 @@
-package org.adoptopenjdk.jheappo.objects
-
-/*
- * Copyright (c) 2018 Kirk Pepperdine.
- * Licensed under https://github.com/AdoptOpenJDK/jheappo/blob/master/LICENSE
- * Instructions: https://github.com/AdoptOpenJDK/jheappo/wiki
- */
-
-import org.adoptopenjdk.jheappo.heapdump.EncodedChunk
-
-class RootThreadBlock(buffer: EncodedChunk) : HeapObject(buffer) {
-
-    companion object {
-        const val TAG: UByte = 0x06U
-    }
-
-    private val threadSerialNumber: UInt = buffer.extractU4()
-
-    override fun toString(): String {
-        return "Root Sticky Class : $id : $threadSerialNumber"
-    }
-}
